id-Attribute in collection-entities
" An attribute "id" belonging to the entity is needed in order for the sets of answers to be saved properly." added to the description.

Is my assumption correct? Sets of answers only worked properly for me after I added the "id-"attribute but I could not find anything related in the documentation.